pyproject.toml
[build-system]
requires = ["setuptools>=60", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "maze_game"
version = "0.0.1"
description = "RL Implementation Playground on Maze Game"
readme = "README.md"
authors = ["berkaytrhn berkaytrhn@gmail.com"]
license = "MIT"
keywords = ["reinforcement learning", "implementations", "game"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
requires-python = ">=3.11.8"



[tool.poetry.urls]
"Source" = "https://github.com/berkaytrhn/maze-game"



[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.poetry.dependencies]
asgiref="3.8.1"
asttokens="3.0.0"
cloudpickle="3.1.0"
colorama="0.4.6"
contourpy="1.3.1"
cycler="0.12.1"
decorator="5.1.1"
Django="5.1.4"
executing="2.1.0"
Farama-Notifications="0.0.4"
fonttools="4.55.0"
gymnasium="1.0.0"
image="1.5.33"
imageio="2.36.1"
imageio-ffmpeg="0.5.1"
ipython="8.30.0"
jedi="0.19.2"
kiwisolver="1.4.7"
matplotlib="3.9.2"
matplotlib-inline="0.1.7"
numpy="2.1.3"
packaging="24.2"
parso="0.8.4"
pillow="11.0.0"
prompt_toolkit="3.0.48"
psutil="6.1.0"
pure_eval="0.2.3"
pygame="2.6.1"
Pygments="2.18.0"
pyparsing="3.2.0"
python-dateutil="2.9.0.post0"
six="1.16.0"
sqlparse="0.5.2"
stack-data="0.6.3"
tqdm="4.67.1"
traitlets="5.14.3"
typing_extensions="4.12.2"
tzdata="2024.2"
wcwidth="0.2.13"

